requirements_api.txt
# Existing dependencies from your project
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
scipy>=1.10.0
torch>=2.0.0
transformers>=4.30.0

# API-specific dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0